Fix compiler warnings for GLib deprecated functions and type mismatches
Replace deprecated GLib functions:
  g_memdup -> g_memdup2 (search.c, uit_fl.c, db.c)
  g_pattern_match_string -> g_pattern_spec_match_string (commands.c)

Fix const qualifier warnings:
  strchr() results cast to (char *) (commands.c, ui_logwindow.c, vars.c)

Fix format string type mismatch:
  cast to (guint64) for G_GUINT64_FORMAT (fl_load.c)

+++ src/fl_load.c
@@ -286,7 +286,7 @@ static fl_list_t *fl_load_parse(int fd, bz_stream *bzs, gboolean local, GError *
       fl_load_token(x, r, err);
     }
     if(*err) {
-      g_prefix_error(err, "Line %"G_GUINT32_FORMAT":%"G_GUINT64_FORMAT": ", x->x.line, x->x.byte);
+      g_prefix_error(err, "Line %"G_GUINT32_FORMAT":%"G_GUINT64_FORMAT": ", x->x.line, (guint64)x->x.byte);
       break;
     }
   }
